refactor docekr compose filenames
