+++ core/server/api_container/server/startosis_engine/docker_compose_transpiler/docker_compose_transpiler_test.go
